MWPW-152988 - populateLocalizedTextFromListItems moved to utils
