+++ eds/blocks/knowledge-base-overview/knowledge-base-overview.js
@@ -1,4 +1,4 @@
-import { getLibs, replaceText, getConfig } from './../../scripts/utils.js';
+import { getLibs, replaceText, getConfig, populateLocalizedTextFromListItems } from './../../scripts/utils.js';
 import { KnowledgeBaseOverview } from './KnowledgeBaseOverview.js';
 
 function declareKnowledgeBaseOverview() {
@@ -40,12 +40,7 @@ export default async function init(el) {
     '{{search}}': 'Search',
   };
 
-  const liList = Array.from(el.querySelectorAll('li'));
-  liList.forEach(liEl => {
-    let liContent = liEl.innerText.trim().toLowerCase().replace(/ /g, '-');
-    if (liContent.endsWith('_default')) liContent = liContent.slice(0, -8);
-    localizedText[`{{${liContent}}}`] = liContent;
-  });
+  populateLocalizedTextFromListItems(el, localizedText);
 
   const deps = await Promise.all([
     localizationPromises(localizedText, config),
