+++ eds/scripts/utils.js
@@ -55,3 +55,12 @@ export { createTag, localizeLink, getConfig };
 
 const { replaceText } = await import(`${miloLibs}/features/placeholders.js`);
 export { replaceText };
+
+export function populateLocalizedTextFromListItems(el, localizedText) {
+  const liList = Array.from(el.querySelectorAll('li'));
+  liList.forEach(liEl => {
+    let liContent = liEl.innerText.trim().toLowerCase().replace(/ /g, '-');
+    if (liContent.endsWith('_default')) liContent = liContent.slice(0, -8);
+    localizedText[`{{${liContent}}}`] = liContent;
+  });
+}
